pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "office-agent"
version = "0.1.0"
description = "Local offline AI agent for Microsoft Office document generation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "click>=8.1",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "python-pptx>=0.6",
    "jsonschema>=4.21",
    "requests>=2.31",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
office_agent = "office_agent.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["office_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
